requirements.txt
# Run requirements for GABM
# Install with: pip install -r requirements.txt

# LLM API dependencies
openai==2.21.0
#anthropic==0.25.6
google-generativeai==0.4.1
httpx==0.28.1
deepseek==1.0.0

# Dependencies for Apertus
transformers==5.1.0
torch==2.10.0
